% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_by_Olga.R
\name{run_linreg}
\alias{run_linreg}
\title{Quickresults}
\usage{
run_linreg(df, xvars, y, terms)
}
\arguments{
\item{df}{mids object multiple imputed dataset containing exposure, outcome and covariates}

\item{xvars}{independent variables}

\item{y}{dependent variable}

\item{terms}{covariates/confounders}
}
\value{
matrix including estimate, beta, se, t, p and FDR-corrected p
}
\description{
Quickresults
}
